+++ src/index.ts
@@ -1,5 +1,5 @@
 import * as apollo from "./apollo"
 export { apollo }
-export { default as query } from "./query"
+export * from "./Query"
 export { default as mutation } from "./mutation"
 export * from "./types"