+++ src/types.ts
@@ -1,13 +1,8 @@
 import { GraphQLError } from "graphql"
 import { FetchMoreOptions, FetchMoreQueryOptions } from "apollo-client"
 
-export interface QueryAttributes<Data = {}, Variables = {}> {
-  data: Data | null | undefined
-  errors: GraphQLError[] | null | undefined
-  variables: Variables
-  loading: boolean
-  fetchMore: (options: FetchMoreOptions & FetchMoreQueryOptions) => void
-  refetch: () => void
+export interface OperationVariables {
+  [key: string]: any
 }
 
 export interface MutationAttributes<Data = {}, Variables = {}> {
@@ -16,4 +11,20 @@ export interface MutationAttributes<Data = {}, Variables = {}> {
   loading: boolean
   called: boolean
   execute: (data: { variables?: Variables }) => void
+}
+
+// Imported from react-apollo. Replace with original if it is ported back to Apollo Client
+export interface FetchMoreOptions<TData, TVariables> {
+  updateQuery: (
+    previousQueryResult: TData,
+    options: {
+      fetchMoreResult?: TData
+      variables: TVariables
+    }
+  ) => TData
+}
+
+// Imported from react-apollo. Replace with original if it is ported back to Apollo Client
+export interface FetchMoreQueryOptions<TVariables, K extends keyof TVariables> {
+  variables: Pick<TVariables, K>
 }