+++ src/util/compact.ts
@@ -1,0 +1,11 @@
+export default function compact(obj: any) {
+  return Object.keys(obj).reduce(
+    (acc, key) => {
+      if (obj[key] !== undefined) {
+        acc[key] = obj[key]
+      }
+      return acc
+    },
+    {} as any
+  )
+}