+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
 ## [Unreleased]
+### Changed
+- Rewrite `<Query/>` as react-apollo.
+
 ### Added
 - Support `fetchMore` and `notificyNetworkStatusChange`